+++ Assets/Script/Character/Hero/Hero.cs
@@ -6,7 +6,7 @@ public abstract class Hero : Character
 
     public Item[] ItemEquip { get; set; }
 
-    public void InitializeHero(int startHealth, int startDamage, int startMoveSpeed, int startAtkCD, int startAtkRange)
+    public void InitializeHero(int startHealth, int startDamage, int startMoveSpeed, float startAtkCD, int startAtkRange)
     {
         base.InitializeCharacter(startHealth, startDamage, startMoveSpeed, startAtkCD, startAtkRange);
